+++ NEWS.md
@@ -11,7 +11,7 @@
   which specific actions are taken during a `clean()` operation.
 
 * `renv` no longer performs an automatic snapshot after a call to
-  `renv::snapshot()`. (#651)
+  `renv::status()`. (#651)
 
 * Fixed an issue where attempts to transform RSPM repository URLs could
   fail if the copy of R was installed without a local `CRAN_mirrors.csv`