+++ NEWS.md
@@ -1,5 +1,5 @@
 
-# renv 0.14.0 (UNRELEASED)
+# renv 0.13.2
 
 * `renv::install("user/repo/subdir with spaces")` now works as expected. (#694)